fix(inaccuracies): Date formats & point baseline
The previously accepted date format was DD MMM YYYY HH:MM.
The new format is now YYYY-MM-DD HH:MM:SS.
The reason for this change is to follow ISO 8601.

Point baseline was mistakenly 750 instead of 900.
The minimum points a player can have is 750, previously there
was no guard to prevent points from going below that threshold.